+++ src/Components/Table/TableHeaderCell.tsx
@@ -46,10 +46,10 @@ export class TableHeaderCell extends Component<Props> {
     const {horizontalAlignment, verticalAlignment, width, style} = this.props
 
     return {
-      ...style,
       textAlign: horizontalAlignment,
       verticalAlign: verticalAlignment,
       width,
+      ...style,
     }
   }