+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ### 0.0.25 (Unreleased)
 
+- [#221](https://github.com/influxdata/clockface/pull/221): [Breaking] Rename `ComponentSpacer` to `FlexBox` and `ComponentSpacerFlexChild` to `FlexBoxChild`
 - [#220](https://github.com/influxdata/clockface/pull/220): Ensure all `ConfirmationButton` elements receive a `testID` attribute
 - [#219](https://github.com/influxdata/clockface/pull/219): Add `style` prop to all components
+- [#219](https://github.com/influxdata/clockface/pull/219): [Breaking] remove `color` prop from `Icon`
 - [#219](https://github.com/influxdata/clockface/pull/219): Ensure `className` prop in `TextArea` is being implemented correctly
 
 ### 0.0.24